CR: block relocation by black-holing file chunks
